Settings: Fix crash in fingerprint enrollment
Like commit() but allows the commit to be executed after an activity's state is saved. This is dangerous because the commit can be lost if the activity needs to later be restored from its state, so this should only be used for cases where it is okay for the UI state to change unexpectedly on the user.
https://developer.android.com/reference/android/app/FragmentTransaction.html

Change-Id: I6334b64531f927be8d1b71dcdbd90f109c936794